会议人员列表-删除功能

+++ guns-vip-main/src/main/java/cn/stylefeng/guns/meetRegister/controller/MeetMemberController.java
@@ -234,13 +234,14 @@ public class MeetMemberController extends BaseController {
     @ResponseBody
     public ResponseData delete(MeetMemberParam meetMemberParam,HttpServletRequest request) {
         String thesisIdStr = request.getParameter("thesisId");
-        //同时删除论文
-        long thesisId = Long.parseLong(thesisIdStr);
-        ThesisParam thesisParam = new ThesisParam();
-        thesisParam.setThesisId(thesisId);
-
+        if(thesisIdStr != null && !("0").equals(thesisIdStr)){
+            //同时删除论文
+            long thesisId = Long.parseLong(thesisIdStr);
+            ThesisParam thesisParam = new ThesisParam();
+            thesisParam.setThesisId(thesisId);
+            this.thesisService.delete(thesisParam);
+        }
         this.meetMemberService.delete(meetMemberParam);
-        this.thesisService.delete(thesisParam);
         return ResponseData.success();
     }
